+++ control/app/Http/Controllers/NewsController.php
@@ -28,8 +28,13 @@ class NewsController extends Controller
      */
     public function index()
     {
-        $news = News::all();
-        return view('appviews.newshow',['news'=>$news]);
+        $logued_user = Auth::user();
+        if($logued_user->usrc_admin || $logued_user->can('see.news')){
+            $news = News::all();
+            return view('appviews.newshow',['news'=>$news]);
+        }else{
+            return view('errors.403');
+        }
     }
 
     /**
@@ -39,7 +44,12 @@ class NewsController extends Controller
      */
     public function create()
     {
-        return view('appviews.newscreate');
+        $logued_user = Auth::user();
+        if($logued_user->usrc_admin || $logued_user->can('create.news')){
+            return view('appviews.newscreate');
+        }else{
+            return view('errors.403');
+        }
     }
 
     /**
@@ -51,13 +61,17 @@ class NewsController extends Controller
     public function store(Request $request)
     {
         $logued_user = Auth::user();
-        $alldata = $request->all();
-        $news = new News($alldata);
-        $news->save();
-        $fmessage = 'Se ha creado la noticia: '.$alldata['tittle'];
-        \Session::flash('message',$fmessage);
-        $this->registeredBinnacle($request->all(), 'store', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
-        return redirect()->route('news.index');
+        if($logued_user->usrc_admin || $logued_user->can('create.news')){
+            $alldata = $request->all();
+            $news = new News($alldata);
+            $news->save();
+            $fmessage = 'Se ha creado la noticia: '.$alldata['tittle'];
+            \Session::flash('message',$fmessage);
+            $this->registeredBinnacle($request->all(), 'store', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
+            return redirect()->route('news.index');
+        }else{
+            return view('errors.403');
+        }
     }
 
     /**
@@ -79,8 +93,13 @@ class NewsController extends Controller
      */
     public function edit($id)
     {
-        $news = News::findOrFail($id);
-        return view('appviews.newsedit',['news'=>$news]);
+        $logued_user = Auth::user();
+        if($logued_user->usrc_admin || $logued_user->can('edit.news')){
+            $news = News::findOrFail($id);
+            return view('appviews.newsedit',['news'=>$news]);
+        }else{
+            return view('errors.403');
+        }
     }
 
     /**
@@ -93,17 +112,21 @@ class NewsController extends Controller
     public function update(Request $request, $id)
     {
         $logued_user = Auth::user();
-        $alldata = $request->all();
-        $news = News::findOrFail($id);
-        $news->tittle = $alldata['tittle'];
-        $news->pdate = $alldata['pdate'];
-        $news->description = $alldata['description'];
-        $news->nactivo = $alldata['nactivo'];
-        $news->save();
-        $fmessage = 'Se ha actualizado la noticia: '.$alldata['tittle'];
-        \Session::flash('message',$fmessage);
-        $this->registeredBinnacle($request->all(), 'update', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
-        return redirect()->route('news.index');
+        if($logued_user->usrc_admin || $logued_user->can('edit.news')){
+            $alldata = $request->all();
+            $news = News::findOrFail($id);
+            $news->tittle = $alldata['tittle'];
+            $news->pdate = $alldata['pdate'];
+            $news->description = $alldata['description'];
+            $news->nactivo = $alldata['nactivo'];
+            $news->save();
+            $fmessage = 'Se ha actualizado la noticia: '.$alldata['tittle'];
+            \Session::flash('message',$fmessage);
+            $this->registeredBinnacle($request->all(), 'update', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
+            return redirect()->route('news.index');
+        }else{
+            return view('errors.403');
+        }
     }
 
     /**
@@ -115,14 +138,18 @@ class NewsController extends Controller
     public function destroy($id,Request $request)
     {
         $logued_user = Auth::user();
-        if (isset($id)){
-            $news = News::findOrFail($id);
-            $fmessage = 'Se ha eliminado la noticia: '.$news->tittle;
-            \Session::flash('message',$fmessage);
-            $this->registeredBinnacle($request->all(), 'destroy', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
-            $news->delete();
+        if($logued_user->usrc_admin || $logued_user->can('delete.news')){
+            if (isset($id)){
+                $news = News::findOrFail($id);
+                $fmessage = 'Se ha eliminado la noticia: '.$news->tittle;
+                \Session::flash('message',$fmessage);
+                $this->registeredBinnacle($request->all(), 'destroy', $fmessage, $logued_user ? $logued_user->id : '', $logued_user ? $logued_user->name : '');
+                $news->delete();
 
+            }
+            return redirect()->route('news.index');
+        }else{
+            return view('errors.403');
         }
-        return redirect()->route('news.index');
     }
 }
